pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skyproject"
version = "0.2.0"
description = "Self-evolving AI development system — PM AI plans, IrgatAI builds, both evolve continuously"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Burak Yilmaz", email = "brkylmzco@gmail.com"},
]
keywords = ["ai", "autonomous", "self-evolving", "code-generation", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai>=1.0",
    "anthropic>=0.20",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "aiofiles>=23.0",
    "orjson>=3.9",
    "chromadb>=1.0",
    "requests>=2.31",
    "networkx>=3.0",
    "tenacity>=8.0",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.urls]
Homepage = "https://github.com/brkylmzco/skyproject"
Repository = "https://github.com/brkylmzco/skyproject"
Issues = "https://github.com/brkylmzco/skyproject/issues"

[project.scripts]
skyproject = "skyproject.cli:main"

[tool.setuptools.packages.find]
include = ["skyproject*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "skyproject"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
